added test for orca molecule parse #169
